+++ build_runner_core/test/asset/finalized_reader_test.dart
@@ -49,7 +49,10 @@ void main() {
         [],
         computeDigest(AssetId('a', 'lib/b.txt'), 'b'),
       );
-      deleted.mutate.deletedBy.add(deleted.id.addExtension('.post_anchor.1'));
+
+      deleted = deleted.rebuild(
+        (b) => b..deletedBy.add(deleted.id.addExtension('.post_anchor.1')),
+      );
 
       graph
         ..add(notDeleted)
@@ -68,7 +71,7 @@ void main() {
       var id = AssetId('a', 'web/a.txt');
       var node = AssetNode.generated(
         id,
-        state: PendingBuildAction.none,
+        pendingBuildAction: PendingBuildAction.none,
         phaseNumber: 0,
         wasOutput: true,
         isFailure: true,