+++ build_runner/CHANGELOG.md
@@ -9,6 +9,7 @@
 - Refactor `FileBasedAssetReader` and `FileBasedAssetWriter` to `ReaderWriter`.
 - Remove `OnDeleteWriter`, add functionality to `ReaderWriter`.
 - Add `NodeType` to `AssetNode`, remove subtypes. Make mutations explicit.
+- Use `built_value` for `AssetNode` and related types.
 
 ## 2.4.15